+++ ios/InteractiveReader/InteractiveReader/Features/Playback/LibraryPlaybackView+Resume.swift
@@ -36,6 +36,9 @@ extension LibraryPlaybackView {
         guard musicOwnership.ownershipState == .appleMusicBed else { return }
         guard lastReaderTransportAction != "play" else { return }
         #if os(tvOS)
+        #if DEBUG
+        e2eTVInteractiveMusicDeferredResumeCount += 1
+        #endif
         resumeAppleMusicBedFromReaderTransportIfNeeded(deferUntilReaderActive: true)
         #else
         musicOwnership.resumeReadingBedForReaderTransport()